+++ src/frontend/apps/standalone_site/src/features/Playlist/features/UpdatePlaylist/components/UpdatePlaylistPage.tsx
@@ -97,6 +97,7 @@ export const UpdatePlaylistPage = () => {
             }}
             submitTitle={intl.formatMessage(messages.saveButtonTitle)}
             isSubmitting={isSubmitting}
+            playlistId={playlistId}
           />
         </Box>
       </WhiteCard>